+++ packages/cli/src/commands/deploy.ts
@@ -35,6 +35,7 @@ export type AppManifest = {
     metadata?: {
         version?: number;
     };
+    experimental?: boolean;
 };
 
 const makeFilesDict = async (glob: string, ignoreGlobs?: string[]) => {
@@ -160,7 +161,13 @@ export const createDeployment = async (
             }
         }
     } catch (error) {
-        Logger.error('The deployment has failed and was aborted due to an error:', error as string);
+        if (typeof error === 'string') {
+            Logger.error('The deployment has failed and was aborted due to an error:', error);
+        } else if (error instanceof Error) {
+            Logger.error('The deployment has failed and was aborted due to an error:', error.message);
+        } else {
+            Logger.error('The deployment has failed and was aborted due to an unknown error.');
+        }
         process.exit(-1);
     }
 };