+++ packages/guideline-blocks-settings/src/components/Link/LinkInput.tsx
@@ -1,7 +1,8 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 import { type AppBridgeBlock } from '@frontify/app-bridge';
-import { ButtonSize, Checkbox, CheckboxState, FormControl, TextInput, TooltipPosition } from '@frontify/fondue';
+import { FormControl } from '@frontify/fondue';
+import { TextInput, Checkbox, Label } from '@frontify/fondue/components';
 
 import { LinkSelector } from './LinkSelector';
 import { isValidUrlOrEmpty as internalIsValidUrlOrEmpty } from './utils';
@@ -11,15 +12,13 @@ type LinkInputProps = {
     info?: string;
     label?: string;
     required?: boolean;
-    newTab?: CheckboxState;
-    openInNewTab?: boolean;
+    newTab?: boolean;
     onUrlChange?: (value: string) => void;
     onToggleTab?: (checked: boolean) => void;
     isValidUrlOrEmpty?: (url: string) => boolean;
     appBridge: AppBridgeBlock;
-    clearable?: boolean;
     placeholder?: string;
-    buttonSize?: ButtonSize;
+    buttonSize?: 'small' | 'medium' | 'large';
     hideInternalLinkButton?: boolean;
 };
 
@@ -28,10 +27,8 @@ export const LinkInput = ({
     onToggleTab,
     isValidUrlOrEmpty,
     appBridge,
-    clearable,
     placeholder,
     newTab,
-    openInNewTab,
     url = '',
     required,
     info,
@@ -40,7 +37,6 @@ export const LinkInput = ({
     hideInternalLinkButton,
 }: LinkInputProps) => {
     const isUrlValid = isValidUrlOrEmpty ? isValidUrlOrEmpty(url) : internalIsValidUrlOrEmpty(url);
-    const checkedState = newTab ?? (openInNewTab ? CheckboxState.Checked : CheckboxState.Unchecked);
     return (
         <div data-test-id="link-input">
             <FormControl
@@ -48,16 +44,15 @@ export const LinkInput = ({
                     children: label,
                     htmlFor: 'url',
                     required,
-                    tooltip: info ? { content: info, position: TooltipPosition.Top } : undefined,
+                    tooltip: info ? { content: info, position: 'top' } : undefined,
                 }}
             >
                 <TextInput
+                    data-test-id="text-input"
                     id="url"
                     value={url}
-                    clearable={clearable}
-                    onChange={onUrlChange}
+                    onChange={(event) => onUrlChange?.(event.target.value)}
                     placeholder={placeholder ?? 'https://example.com'}
-                    focusOnMount
                 />
             </FormControl>
             {!isUrlValid && <div className="tw-text-text-negative tw-mt-1 tw-text-s">Please enter a valid URL.</div>}
@@ -67,7 +62,7 @@ export const LinkInput = ({
                     <LinkSelector
                         url={url}
                         onUrlChange={onUrlChange}
-                        buttonSize={buttonSize ?? ButtonSize.Medium}
+                        buttonSize={buttonSize ?? 'medium'}
                         getAllDocuments={() => appBridge.getAllDocuments()}
                         getDocumentPagesByDocumentId={(documentId) =>
                             appBridge.getDocumentPagesByDocumentId(documentId)
@@ -79,8 +74,11 @@ export const LinkInput = ({
                 </div>
             )}
 
-            <div className="tw-mt-3">
-                <Checkbox value="new-tab" label="Open in new tab" state={checkedState} onChange={onToggleTab} />
+            <div className="tw-mt-3 tw-flex tw-items-center tw-gap-1.5">
+                <Checkbox id="new-tab" value={newTab} onChange={() => onToggleTab?.(!newTab)} />
+                <Label id="new-tab-label" htmlFor="new-tab">
+                    Open in new tab
+                </Label>
             </div>
         </div>
     );